+++ changes.md
@@ -1,5 +1,10 @@
 # Latest Changes
 
+## dev
+- Update fornax-labextension to 0.1.22: Change the label Introduction to Tutorials.
+- Add s3archive script to help efficiently archive data to ~/s3-storage.
+- Bug fixes and enhancements:
+
 ## 26.0814
 - Hotfix for the missing /opt/support-data symlink.